backend/pyproject.toml
[tool.poetry]
name = "kace-mobile"
version = "0.1.0"
description = "Mobile School Management System"
authors = ["Songtaa <songtaa9@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["standard"], version = "^0.115.5"}
uvicorn = "^0.30.6"
python-jose = "^3.3.0"
pydantic-settings = "^2.4.0"
pillow = "^10.4.0"
python-levenshtein = "^0.25.1"
ruff = "^0.6.1"
sentry-sdk = "^2.13.0"
sqlalchemy = "^2.0.34"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
inflect = "^7.4.0"
pydantic = {extras = ["email"], version = "^2.9.1"}
requests = "^2.32.3"
fastapi-mail = "^1.4.1"
alembic = "^1.13.3"
passlib = "^1.7.4"
pyjwt = "^2.9.0"
sqlmodel = "^0.0.22"
aioredis = "^2.0.1"
rq = "^2.0.0"
asyncpg = "^0.30.0"
pendulum = "^3.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
